rename interpolateRgb -> interpolateRGB, add color functions
